% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_style.R
\docType{data}
\name{OI_COLORS}
\alias{OI_COLORS}
\title{The OI color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
OI_COLORS
}
\description{
The OI color palette
}
\keyword{datasets}
