% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_style.R
\name{set_oi_palette}
\alias{set_oi_palette}
\title{Add oi theme to ggplot chart}
\usage{
set_oi_palette()
}
\description{
This function allows you to add the oi theme to your ggplotgraphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
oi_style()
Set global color options
}
\keyword{oi_style}
