% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oi_style.R
\name{theme_oi}
\alias{theme_oi}
\title{ggplot clone of the OI Stata scheme}
\usage{
theme_oi(high_contrast = FALSE, font = "Helvetica")
}
\description{
ggplot clone of the OI Stata scheme
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
  geom_line(colour = "#007f7f", size = 1) +
  geom_hline(yintercept = 0, size = 1, colour = "#333333") +
  oi_style()
}
